Pod k8s 관련 오류 수정
- 타입 수정

+++ src/main/java/kr/co/strato/adapter/k8s/pod/service/PodAdapterService.java
@@ -128,7 +128,7 @@ public class PodAdapterService {
         }
     }
 
-    public String getYaml(Integer clusterId, String namespaceName, String podName){
+    public String getYaml(Long clusterId, String namespaceName, String podName){
         try{
             String yaml = inNamespaceProxy.getResourceYaml(ResourceType.pod.get(), clusterId, namespaceName, podName);
